% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.R
\name{AIC}
\alias{AIC}
\title{Akaike Information Criterion}
\usage{
AIC(LogLik, k)
}
\arguments{
\item{LogLik}{The loglikelhood or a vector of loglikelihood values}

\item{k}{The number of parameters}
}
\value{
The AIC value
}
\description{
Akaike Information Criterion
}
\author{
Rampal S. Etienne
}
