% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICc.R
\name{AICc}
\alias{AICc}
\title{Corrected Akaike Information Criterion}
\usage{
AICc(LogLik, k, n)
}
\arguments{
\item{LogLik}{The loglikelhood or a vector of loglikelihood values}

\item{k}{The number of parameters}

\item{n}{The number of observations (sample size)}
}
\value{
The AICc value
}
\description{
Corrected Akaike Information Criterion
}
\author{
Rampal S. Etienne
}
