% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICweights.R
\name{ICweights}
\alias{ICweights}
\title{Information Criterion Weights}
\usage{
ICweights(IC)
}
\arguments{
\item{IC}{The value(s) of the Information Criterion}
}
\value{
The IC weights for a set of models for which loglikelihoods are provided.
}
\description{
Information Criterion Weights
}
\author{
Rampal S. Etienne
}
