% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isar_alpine_lakes.R
\name{isar_alpine_lakes}
\alias{isar_alpine_lakes}
\title{Estimates parameters of various island species-area relationships (isars)}
\usage{
isar_alpine_lakes(
  type_of_richness = "D1.OC",
  trial_settings = c(1, 0.1),
  working_directory = getwd()
)
}
\arguments{
\item{type_of_richness}{Type of richness that is used. Choice is one of
"D1.OC", "D2.OC_Int", "D3.OC_Ext", "D6.CC".
D1.OC: the diversity/richness from the Original Community
D2.OC_Int: Original community richness plus introduced species
D3.OC_Ext: Original community minus extinct species
D6.CC: Contemporary Community, that is, OC plus Introduced species minus
Extinct species.}

\item{trial_settings}{Number of trials in optimization, and standard deviation
of the normal distribution used to generate the trials (initial parameter sets)}

\item{working_directory}{Name of working directory (as a string) where results
result files will be put. Default is current working directory.}
}
\value{
The estimated parameters for each isar, the loglikelihood, the AICc
and the AICc weights.
}
\description{
Uses maximum likelihood to estimate parameters of various species-
area relationships that have been proposed in the literature. A pdf is
produced that contains the plots of the 4 best-fitting isars. The data is also
written to file. These files are put in the working directory that can be
specified.
}
\details{
Estimation of parameters of isars for fish in   alpine lakes
}
\author{
Rampal S. Etienne
}
