% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isar_loglik.R
\name{isar_loglik}
\alias{isar_loglik}
\title{Computes the loglikelihood of a parameter set of a specified island
species-area relationships (isars) for a given data set of values of
area and richness.}
\usage{
isar_loglik(f_isar, pars, area, obs_richness)
}
\arguments{
\item{f_isar}{Function for island species-area relationship}

\item{pars}{A numeric vector containing the model parameters:
\itemize{
  \item{\code{pars[1]}: }
  \item{\code{pars[2]}: }
}}

\item{area}{Vector of values of areas for different islands}

\item{obs_richness}{vector of values of observed richness on islands. Should
be the same order as in area.}
}
\value{
The loglikelihood.
}
\description{
Computes the logarithm of the likelihood of a model for the
island species-area with a given set of parameters for a given data set with
values of area and richness for several islands. It assumes that the island
species-area relationship provides the expected value of richness, and the
error distribution is Poisson.
}
\details{
Likelihood of isar model
}
\author{
Rampal S. Etienne
}
