% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isar_ML.R
\name{isar_ML}
\alias{isar_ML}
\title{Maximum Likelihood of isar models}
\usage{
isar_ML(
  f_isar,
  initparsopt,
  parsfix,
  idparsopt,
  idparsfix,
  area,
  obs_richness,
  optimmethod = "simplex",
  num_cycles = 1,
  tol = c(1e-06, 1e-06, 1e-06),
  maxiter = 10000 * round((1.25)^length(idparsopt)),
  trial_settings = c(n = 10, sd = 0.1)
)
}
\arguments{
\item{f_isar}{Function for island species-area relationship}

\item{initparsopt}{The initial values of the parameters that must be
optimized; they are all positive.}

\item{parsfix}{The values of the parameters that should not be optimized.}

\item{idparsopt}{The ids of the parameters that must be optimized.}

\item{idparsfix}{The ids of the parameters that should not be optimized,}

\item{area}{vector of values of areas for different islands}

\item{obs_richness}{vector of values of observed richness on islands. Should
be the same order as in area.}

\item{optimmethod}{Method used in likelihood optimization. For example
\code{simplex} or \code{subplex}. See \code{\link[DDD]{optimizer}()} for
further details.}

\item{num_cycles}{The number of cycles the optimizer will go through.
Default is 1.}

\item{tol}{Sets the tolerances in the optimization. Consists of: \cr reltolx
= relative tolerance of parameter values in optimization \cr reltolf =
relative tolerance of function value in optimization \cr abstolx = absolute
tolerance of parameter values in optimization.}

\item{maxiter}{Sets the maximum number of iterations in the optimization.}

\item{trial_settings}{number of trials in optimization, and standard deviation
of the normal distribution used to generate the trials (initial parameter sets)}
}
\value{
The maximum likelhood parameters and corresponding likelihood and AICc value.
}
\description{
Maximum Likelihood of isar models
}
\author{
Rampal S. Etienne
}
